requirements.txt
python-dotenv
rich
colorama
openai
graphiti-core[falkordb]
openai
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
apscheduler>=3.10.0
langfuse
